+++ src/components/Jobs.jsx
@@ -1,7 +1,11 @@
-import { useState } from 'react';
-import Job from './Job';
 import './Jobs.css';
+import { useState } from 'react';
 import Header from './Header';
+import Button from '../components/utility/Button';
+import JobItem from './JobItem';
+import JobHeader from './JobHeader';
+
+// Images
 import tankIcon from '../images/tank.png';
 import healerIcon from '../images/healer.png';
 import meleeIcon from '../images/melee.png';
@@ -9,7 +13,6 @@ import rangedIcon from '../images/ranged.png';
 import magicIcon from '../images/magic.png';
 import handIcon from '../images/hand.png';
 import landIcon from '../images/land.png';
-import Button from '../components/utility/Button';
 
 
 const Jobs = (props) => {
@@ -19,8 +22,10 @@ const Jobs = (props) => {
     const handLandJobs = props.jobs.slice(20);
 
     return (
-        <div className="jobs section">
+        <div className="section">
+
             <Header name="Jobs" />
+
             <div className="row gap-lg">
                 <Button 
                     content="War / Magic" 
@@ -33,41 +38,44 @@ const Jobs = (props) => {
                     onClick={() => setDisplayJob(false)}
                 />
             </div>
-            <div className="panel">
-                <div className={"jobs--war-magic panel__content panel__content--" + props.type + " " + (displayJob ? "" : "disabled")}>
-                    <Header name="Tank" isMinor={true} image={tankIcon} />
-                    <Header name="Healer" isMinor={true} image={healerIcon} />
-                    <Header name="Melee" isMinor={true} image={meleeIcon} />
-                    <Header name="Ranged" isMinor={true} image={rangedIcon} />
-                    <Header name="Magic" isMinor={true} image={magicIcon} />
 
-                    {warMagicJobs.map(job => 
-                        <Job
-                        key={job.Job.ID}
-                        icon={job.Job.Icon}
-                        name={job.Job.Name}
-                        level={job.Level}
-                        exp={[job.ExpLevel, job.ExpLevelMax]}
-                        hasLink={true}
-                        />
-                    )}
-                </div>
-                <div className={"jobs--hand-land panel__content panel__content--" + props.type + " " + (!displayJob ? "" : "disabled")}>
-                    
-                    <Header name="Hand" isMinor={true} image={handIcon} />
-                    <Header name="Land" isMinor={true} image={landIcon} />
-                    {handLandJobs.map(job => 
-                        <Job
-                        key={job.Job.ID}
-                        icon={job.Job.Icon}
-                        name={job.Job.Name}
-                        level={job.Level}
-                        exp={[job.ExpLevel, job.ExpLevelMax]}
-                        />
-                    )}
-                </div>
+            <div className={"jobs__collection war-magic" + (displayJob ? "" : " disabled")}>
+
+                <JobHeader name="Tank" icon={tankIcon} />
+                <JobHeader name="Healer" icon={healerIcon} />
+                <JobHeader name="Melee" icon={meleeIcon} />
+                <JobHeader name="Ranged" icon={rangedIcon} />
+                <JobHeader name="Magic" icon={magicIcon} />
+
+                {warMagicJobs.map(job => 
+                    <JobItem
+                    key={job.Job.ID}
+                    icon={job.Job.Icon}
+                    name={job.Job.Name}
+                    level={job.Level}
+                    exp={[job.ExpLevel, job.ExpLevelMax]}
+                    hasLink={true}
+                    />
+                )}
+
             </div>
-        </div>  
+            <div className={"jobs__collection hand-land" + (!displayJob ? "" : " disabled")}>
+
+                <JobHeader name="Hand" icon={handIcon} />
+                <JobHeader name="Land" icon={landIcon} />
+                
+                {handLandJobs.map(job => 
+                    <JobItem
+                    key={job.Job.ID}
+                    icon={job.Job.Icon}
+                    name={job.Job.Name}
+                    level={job.Level}
+                    exp={[job.ExpLevel, job.ExpLevelMax]}
+                    />
+                )}
+
+            </div>
+        </div>
     );
 }