+++ src/components/utility/Divider.jsx
@@ -1,0 +1,6 @@
+import './Divider.css';
+const Divider = () => {
+    return (<div className="divider" />);
+}
+
+export default Divider;